requirements.txt
# Core Django
Django==5.1.3
djangorestframework==3.15.2
django-cors-headers==4.4.0

# Database
dj-database-url==2.2.0
psycopg2-binary==2.9.9

# Environment & Configuration
python-decouple==3.8

# Security & Authentication
django-allauth==0.57.0
cryptography==41.0.7

# File Handling
Pillow==10.1.0
django-storages==1.14.2

# API & Serialization
djangorestframework-simplejwt==5.3.0
django-filter==23.5

# Caching & Performance
redis==5.0.1
django-redis==5.4.0

# Development Tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0

# Code Quality
flake8==6.1.0
black==23.11.0
isort==5.12.0

# Monitoring & Logging
sentry-sdk==1.38.0

# Utilities
python-dateutil==2.8.2
requests==2.31.0
celery==5.3.4

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0